+++ Sonar.UserProfile.ApiClient/Interfaces/IApiClientUser.cs
@@ -1,12 +1,10 @@
 ﻿using Sonar.UserProfile.Web.Controllers.Users.Dto;
 
-namespace Sonar.UserProfile.ApiClient;
+namespace Sonar.UserProfile.ApiClient.Interfaces;
 
-public interface IApiClient
+public interface IApiClientUser
 {
     Task<string> RegisterAsync(UserRegisterDto userRegisterDto, CancellationToken cancellationToken);
     Task<string> LoginAsync(UserLoginDto userLoginDto, CancellationToken cancellationToken);
     Task<UserGetDto> GetAsync(string token, CancellationToken cancellationToken);
-    Task AddFriendAsync(string token, string friendEmail, CancellationToken cancellationToken);
-    Task<IReadOnlyList<UserGetDto>> GetFriendsAsync(string token, CancellationToken cancellationToken);
 }
